Fix: rename action creator file to amazingComponent
