+++ app/src/actions/amazingComponent.js
