--- app/src/constants/index.js
@@ -1,12 +1,0 @@
-// Add constants that you use in your app here.
-// Example:
-/*
-const MyConstants = {
-  // Constants for reducer actions
-  ADD_FOO_TO_BAR = "ADD_FOO_TO_BAR",
-  ADD_BAR_TO_FOO = "ADD_BAR_TO_FOO",
-
-  // Constants used in your app:
-  NAV_BAR_TITLE: "React Simple Boilerplate"
-};
-*/
